+++ src/Module_Nodes/generate_shader_code.cpp
@@ -62,6 +62,9 @@ auto generate_shader_code(
 
 uniform float     _time;
 uniform float     _height;
+uniform float     _audio_volume;
+uniform sampler1D _audio_spectrum;
+uniform sampler1D _audio_waveform;
 uniform mat3      _camera2D;
 uniform mat3      _camera2D_inverse;
 uniform sampler2D _previous_frame_texture;